Move utils files to components folder
