+++ components/game/utils/findAffectedRowOrColumn.ts
