+++ components/game/utils/findRoworColumnPosition.ts
